+++ CHANGES.md
@@ -1,4 +1,4 @@
-# Version 0.8.0 (Not Released Yet)
+# Version 0.8.0
 
 ## Features
 - Refactoring SSL generation to work on every request so no domain environment variable is required for SSL generation (#1133)
@@ -6,6 +6,7 @@
 - Added events for functions and executions (#971)
 - Added JWT support (#784)
 - Added ARM support (#726)
+- New awesome image preview features, supports borderRadius, borderColor, borderWidth 
 - Split token & session models to become 2 different internal entities (#922)
 - Added Dart 2.12 as a new Cloud Functions runtime (#989)
 - Added option to disable email/password (#947)
@@ -22,6 +23,8 @@
 - Added a new env var named `_APP_LOCALE` that allow to change the default `en` locale value (#1056)
 - Updated all the console bottom control to be consistent. Dropped the `+` icon (#1062)
 - Added Response Models for Documents and Preferences (#1075, #1102)
+- Added new endpoint to update team membership roles (#1142)
+- Removed DB connection from webhooks worker for improved performance (#1150)
 
 ## Bugs
 
@@ -31,6 +34,7 @@
 - Bugs in the Teams API:
   - Fixed incorrect audit worker event names (#1143)
   - Increased limit of memberships fetched in `createTeamMembership` to 2000 (#1143)
+  - Fixed exception thrown when SSL certificate is already stored in the database (#1151)
 
 ## Breaking Changes (Read before upgrading!)